+++ src/main/java/com/danahub/zipitda/store/dto/ProductRequestDto.java
@@ -2,6 +2,8 @@ package com.danahub.zipitda.store.dto;
 
 import com.danahub.zipitda.store.domain.Category;
 import jakarta.validation.constraints.*;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ProductRequestDto(
@@ -12,8 +14,8 @@ public record ProductRequestDto(
         @NotBlank(message = "카테고리를 입력하세요.")
         Long categoryId,
         @NotNull(message = "가격을 입력하세요.")
-        @Positive Long price,
+        @Positive BigDecimal price,
         @NotNull(message = "재고 수량을 입력하세요.")
-        @PositiveOrZero Long stockQuantity,
+        @PositiveOrZero int stockQuantity,
         List<String> imageUrls
 ) { }
