+++ src/main/java/com/danahub/zipitda/store/dto/ProductResponseDto.java
@@ -1,12 +1,13 @@
 package com.danahub.zipitda.store.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProductResponseDto(
         Long id,
         String name,
         String categoryName,
-        Long price,
-        Long stockQuantity,
+        BigDecimal price,
+        int stockQuantity,
         LocalDateTime createdAt
 ) {}
